requirements.txt
pandas
numpy
torch
transformers
sentence-transformers
Pillow
scikit-learn
tqdm